place bots fairly
The trick is to consider not the final offsets at first, but the
places between the free bytes. Select one for each bot uniformly
at random; a heap is an appropriate data structure to store them.
Pick a random permuation for the order of the bots and iterate over
the positions from start to end. Compute the eventual offsets by
adding to each the amount of space allocated so far.

Thus we obtain any possible placement all over the available memory
without bias.